# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Centros",
        "description": "Documentação da API de Centros, incluindo operações de CRUD e relacionamentos.",
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API principal"
        }
    ],
    "paths": {
...
        "/formadores": {
            "get": {
                "tags": [
                    "Formadores"
                ],
                "summary": "Listar todos os formadores",
                "operationId": "904bf7e097f7a3ec3126089c8af27f43",
                "parameters": [
                    {
                        "name": "busca",
                        "in": "query",
                        "description": "Busca textual em nome, especialidade ou bio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de formadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Formador"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Formadores"
                ],
                "summary": "Criar um novo formador",
                "operationId": "358489cd3e00c6d0c8005b83ad354849",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormadorInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Formador criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formador"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/formadores/{id}": {
            "get": {
                "tags": [
                    "Formadores"
                ],
                "summary": "Buscar formador por ID",
                "operationId": "0f3a04185a92e4dffe1f0bbc63e8d54f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formador encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formador"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formador não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Formadores"
                ],
                "summary": "Atualizar formador",
                "operationId": "47f30ee76b43a8e553be88d470080771",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormadorInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formador atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formador"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Formador não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Formadores"
                ],
                "summary": "Deletar formador",
                "operationId": "c3b3a0df83a21bda91501bae58822409",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formador deletado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Formador não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Formador": {
                "required": [
                    "nome",
                    "contactos"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "especialidade": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    },
                    "foto_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "cursos": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "centros": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "nome": "João Silva",
                    "email": "joao@exemplo.com",
                    "contactos": [
                        "923456789",
                        "912345678"
                    ],
                    "especialidade": "Informática",
                    "bio": "Formador com experiência em cursos de tecnologia.",
                    "foto_url": "https://exemplo.com/foto.jpg",
                    "cursos": [
                        1,
                        2
                    ],
                    "centros": [
                        1
                    ]
                }
            },
            "FormadorInput": {
                "required": [
                    "nome",
                    "contactos"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "especialidade": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    },
                    "foto_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "cursos": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "centros": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "nome": "João Silva",
                    "email": "joao@exemplo.com",
                    "contactos": [
                        "923456789",
                        "912345678"
                    ],
                    "especialidade": "Informática",
                    "bio": "Formador com experiência em cursos de tecnologia.",
                    "foto_url": "https://exemplo.com/foto.jpg",
                    "cursos": [
                        1,
                        2
                    ],
                    "centros": [
                        1
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}